requirements.txt
Flask
Flask-SocketIO
eventlet
requests
websocket-client
gunicorn